AWSapp/requirements.txt
boto3==1.10.45
botocore==1.13.45
pymongo==3.10.0
numpy
pandas
scipy
sklearn
Pillow==6.2.1
mtcnn==0.1.0
matplotlib
tensorflow==1.14
Keras==2.2.5
Flask==1.1.1
Flask-RESTful==0.3.7
Werkzeug==0.16.0
opencv-python